Experimental LCD and debugging support with a Mega2560
This uses an Arduino Mega2560 with a 16x2 LCD shield and some
pushbuttons.  The E&C bus connects to Serial1 (pins 18 and 19) and the
sense is on pin 20.

The limited display area of the LCD isn't good if you're viewing lots of
data that updates frequently.  I've made it echo out the Serial port so
it can be logged.